Correccion retorno factura

+++ pay/crear_pago.php
@@ -29,7 +29,7 @@ try {
     // === Definir datos del pago ===
     // Datos del pago
 	
-	$porcentajeAdicional = ADDITIONAL_PERCENTAGE_PAYMENT;
+	$porcentajeAdicional = (float) ADDITIONAL_PERCENTAGE_PAYMENT;
 	$monto = (float)($cliente['plan_precio'] ?? 0);
 	if ($monto <= 0) $monto = 50000; // valor base si el plan no tiene precio
